+++ src/editor/editor.js
@@ -1,6 +1,6 @@
-import { Annotation, EditorState, Compartment, Facet, EditorSelection, Transaction, Prec } from "@codemirror/state"
+import { Annotation, EditorState, Compartment, Facet, EditorSelection, Transaction, Prec, RangeSet } from "@codemirror/state"
 import { EditorView, keymap as cmKeymap, drawSelection, ViewPlugin, lineNumbers } from "@codemirror/view"
-import { foldGutter, ensureSyntaxTree } from "@codemirror/language"
+import { ensureSyntaxTree, foldState, foldEffect } from "@codemirror/language"
 import { markdown, markdownKeymap } from "@codemirror/lang-markdown"
 import { undo, redo } from "@codemirror/commands"
 
@@ -27,6 +27,7 @@ import { NoteFormat } from "../common/note-format.js"
 import { AUTO_SAVE_INTERVAL } from "../common/constants.js"
 import { useHeynoteStore } from "../stores/heynote-store.js";
 import { useErrorStore } from "../stores/error-store.js";
+import { foldGutterExtension } from "./fold-gutter.js"
 
 
 // Turn off the use of EditContext, since Chrome has a bug (https://issues.chromium.org/issues/351029417) 
@@ -83,9 +84,9 @@ export class HeynoteEditor {
                 heynoteCopyCut(this),
 
                 //minimalSetup,
-                this.lineNumberCompartment.of(showLineNumberGutter ? [lineNumbers(), blockLineNumbers] : []),
+                this.lineNumberCompartment.of(showLineNumberGutter ? blockLineNumbers : []),
                 customSetup, 
-                this.foldGutterCompartment.of(showFoldGutter ? [foldGutter()] : []),
+                this.foldGutterCompartment.of(showFoldGutter ? [foldGutterExtension()] : []),
                 this.closeBracketsCompartment.of(bracketClosing ? [getCloseBracketsExtensions()] : []),
 
                 this.readOnlyCompartment.of([]),
@@ -164,6 +165,13 @@ export class HeynoteEditor {
     getContent() {
         this.note.content = this.view.state.sliceDoc()
         this.note.cursors = this.view.state.selection.toJSON()
+
+        // fold state
+        const foldedRanges = []
+        this.view.state.field(foldState, false)?.between(0, this.view.state.doc.length, (from, to) => {
+            foldedRanges.push({from, to})
+        })
+        this.note.foldedRanges = foldedRanges
         
         const ranges = this.note.cursors.ranges
         if (ranges.length == 1 && ranges[0].anchor == 0 && ranges[0].head == 0) {
@@ -177,7 +185,6 @@ export class HeynoteEditor {
         //console.log("loading content", this.path)
         const content = await window.heynote.buffer.load(this.path)
         this.diskContent = content
-        this.contentLoaded = true
 
         // set up content change listener
         this.onChange = (content) => {
@@ -187,6 +194,7 @@ export class HeynoteEditor {
         window.heynote.buffer.addOnChangeCallback(this.path, this.onChange)
 
         await this.setContent(content)
+        this.contentLoaded = true
     }
 
     setContent(content) {
@@ -230,6 +238,10 @@ export class HeynoteEditor {
                         scrollIntoView: true,
                     })
                 }
+                // set folded ranges
+                this.view.dispatch({
+                    effects: this.note.foldedRanges.map(range => foldEffect.of(range)),
+                })
                 resolve()
             })
         })
@@ -369,13 +381,13 @@ export class HeynoteEditor {
 
     setLineNumberGutter(show) {
         this.view.dispatch({
-            effects: this.lineNumberCompartment.reconfigure(show ? [lineNumbers(), blockLineNumbers] : []),
+            effects: this.lineNumberCompartment.reconfigure(show ? blockLineNumbers : []),
         })
     }
 
     setFoldGutter(show) {
         this.view.dispatch({
-            effects: this.foldGutterCompartment.reconfigure(show ? [foldGutter()] : []),
+            effects: this.foldGutterCompartment.reconfigure(show ? foldGutterExtension : []),
         })
     }