+++ docs/changelog.md
@@ -2,6 +2,18 @@
 
 Here are the most notable changes in each release. For a more detailed list of changes, see the [Github Releases page](https://github.com/heyman/heynote/releases).
 
+## 2.3.0-beta (not released yet)
+
+### Improved support for folding blocks
+
+- Fixed an issue where some types of blocks could not be folded. 
+- The fold state is now stored in the buffer file so that it's persisted between sessions. 
+- Added three new commands for folding and unfolding blocks:
+  - `foldBlock` - Folds the current/selected block(s). Default key binding is `Ctrl-Alt-[` on Windows/Linux and `Cmd-Option-[` on Mac.
+  - `unfoldBlock` - Unfolds the current/selected block(s). Default key binding is `Ctrl-Alt-]` on Windows/Linux and `Cmd-Option-]` on Mac.
+  - `toggleFoldBlock` - Toggles the fold state of the current/selected block. Default key binding is `Ctrl-Alt-.` on Windows/Linux and `Cmd-Option-.` on Mac.
+- Folded blocks now display the first 50 characters of the block content, as well as the number of lines in the block.
+
 ## 2.2.2
 
 - Fix issue when changing settings after having upgraded to Heynote 2.2 from an earlier version